Set invalid measures to missing

invalid_to_na sets invalid measures to missing (NA).

invalid_to_na(pupil, condition, log = FALSE, log_file = NULL)

Arguments

pupil
A numeric vector containing pupil data.

condition
A condition to determine which values should be set to NA.

log
A logical value. Should the action and results be logged?

log_file
A character string specifying the path to the log file.

Examples

# Set 4 to missing in a vector from 1 to 10
invalid_to_na(1:10, 1:10 == 4)

#>  [1]  1  2  3 NA  5  6  7  8  9 10

# Set all pupil sizes with a value of -1 to NA
invalid %>%
  mutate(
    pupil_left = invalid_to_na(pupil_left, pupil_left == -1),
    pupil_right = invalid_to_na(pupil_right, pupil_right == -1)
  )

#> # A tibble: 294 x 6
#>    subject timestamp validity_left validity_right pupil_left pupil_right
#>      <dbl>     <int>         <dbl>          <dbl>      <dbl>       <dbl>
#>  1      48         1             0              0       2.56        2.69
#>  2      48         2             0              0       2.57        2.69
#>  3      48         3             0              0       2.57        2.70
#>  4      48         4             0              0       2.57        2.71
#>  5      48         5             4              0      NA           2.72
#>  6      48         6             0              0       2.58        2.72
#>  7      48         7             4              0      NA           2.72
#>  8      48         8             0              0       2.59        2.72
#>  9      48         9             0              0       2.59        2.72
#> 10      48        10             0              0       2.58        2.71
#> # … with 284 more rows

# Or, set pupil sizes to NA based on a validity score

invalid %>%
  mutate(
    pupil_left = invalid_to_na(pupil_left, validity_left == 4),
    pupil_right = invalid_to_na(pupil_right, validity_right == 4)
  )

#> # A tibble: 294 x 6
#>    subject timestamp validity_left validity_right pupil_left pupil_right
#>      <dbl>     <int>         <dbl>          <dbl>      <dbl>       <dbl>
#>  1      48         1             0              0       2.56        2.69
#>  2      48         2             0              0       2.57        2.69
#>  3      48         3             0              0       2.57        2.70
#>  4      48         4             0              0       2.57        2.71
#>  5      48         5             4              0      NA           2.72
#>  6      48         6             0              0       2.58        2.72
#>  7      48         7             4              0      NA           2.72
#>  8      48         8             0              0       2.59        2.72
#>  9      48         9             0              0       2.59        2.72
#> 10      48        10             0              0       2.58        2.71
#> # … with 284 more rows